+++ src/index.ts
@@ -1,3 +1,5 @@
+"use client";
+
 export * from "./text_editor";
 export * from "./text_editor_controller";
 export * from "./events";